functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
pymongo~=4.7.3
PyJWT~=2.8.0
Werkzeug~=3.0.3
python-dotenv~=1.0.1
sendgrid
opencv-python